fix compilation warnings + update
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@62226 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 6ca5304caa1647103d24b297ee2eaddbcb3736e1

+++ Rec/Calo/CaloMoniDst/src/CaloEMuPIDMon.cpp
@@ -122,7 +122,7 @@ StatusCode CaloEMuPIDMon::execute()
     do {
       if ( !inecal )                                                   break; // ----
 
-      float rdlle = proto->info( LHCb::ProtoParticle::RichDLLe, -9999.);
+      float rdlle = (float) proto->info( LHCb::ProtoParticle::RichDLLe, -9999.);
       if ( rdlle < m_RichDLLe && !m_uncut )                            break; // ----
       if ( inhcal && hcale > m_maxEHcalE && !m_uncut)                  break; // ----
